Extract ConfigurationCacheIncludedBuildScriptChangesIntegrationTest
from ConfigurationCacheIncludedBuildChangesIntegrationTest
